`delete_fragments` Needs To Also Delete Unused Schemas
